requirements.txt
python_version == 3.8.10
streamlit==1.40.1
opencv-python-headless==4.8.1.78
mediapipe==0.10.11
numpy==1.24.3
av==12.3.0
streamlit-webrtc==0.48.2
pillow==10.4.0
scikit-learn==1.2.2
